requirements.txt
# Web 框架
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Milvus
pymilvus>=2.4.0

# 向量提取
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.44.0,<5.0.0
pillow>=10.0.0
sentencepiece>=0.1.99
protobuf>=5.27.2,<6.0.0

# 人脸识别
insightface>=0.7.3
onnxruntime>=1.15.0
onnx>=1.14.0

# OCR (可选，如果后续需要)
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# 机器学习
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
hdbscan>=0.8.33
opencv-python>=4.8.0

# 数据验证
pydantic>=2.0.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0   # PyMySQL caching_sha2_password 认证需要

# 工具
python-dotenv>=1.0.0
certifi>=2024.0.0

